Shatter the optimizer tool into enough files that it's maintainable rather than a wall of code.
